cli: Move 'update-runtime' to 'update runtime'
